Fix max_attempts for StorageUpdaterJob
This sets `max_attempts` for StorageUpdaterJob to 1,
as we prefer having an external event retrigger storage counting
rather than retrying from delayed_job.

The reason behind this is failing counting due to I/O or other
unforeseen issues (e.g. mount not available, .. ).

In this case, recounting soon after will not solve the problem
and only increase load on disk.