Correct height when expanding inplace textareas with autogrow
When using `.-autogrow` on textareas in inplace edit (e.g., description),
the height is not correctly set as the selector is not visible.

Only when starting to EDIT the textarea, the changes are being picked up
and height is set correctly.

This commit thus provides a second event listener to `focusin` events
that cause the textareas to resize properly.