Ensure compatibility of sum-functions with Array callers
this is neccessary to allow building sums for grouped results.
The long-term goal should be to also move the grouping to the SQL side.

Currently grouping will make a query very slow
("normal" with respect to its prior performance,
but "slow" with respect to the performance of the grouping-free sum implementation).