Enforce default isolation level of READ COMMITTED
This sets the session (== ActiveRecord connection) wide default transaction isolation level.

Specific reason: For advisory based locking over journals in MySQL to work correctly,
we need to ensure that the SELECT that determines the maximum version reads from the most recent committed data.
However, by default MySQL would use REPEATABLE READ, which snapshots the database at the first SELECT of a transaction.

Generic reason: Until now we did not expect any specific isolation level,
this might be bad as MySQL and PostgreSQL already use different defaults. It is also possible to reconfigure a DBMS for a different default, so we would either need to require our customers to configure a specified isolation level OR we set one for our own sessions... Now we know for sure that our transactions are using READ COMMITTED unless specified otherwise on the transaction.