Minor corrections
This commit addresses the following minor corrections from review:

1. Remove constant URI:
It breaks karma, and I can't really tell why

2. Avoid using Object.values

Appeared to be supported by chrome, but causes the protractor failure.
This also switches the ordering to binary (unshift, push) as that's all
we need for now.

3. Remove location.hash when jumping to element

`$location.hash` may take precedence over `scrollIntoView` ocassionaly,
thus we forcefully remove the location hash when updating the work
package on a particular field

4. Avoid showing comment success when empty

When opening the comment field, and then submitting on another
field with comment being empty and open, the comment was not submitted,
but a successful comment notification was still sent.

5. Fix issues for IE 10+

* Force min-height to be set on the element, as IE tends to return 0 as
  height when empty

6. Misc.

* Avoid undefined in comment text
* Avoid resizing [msd-elastic] textareas
* Rename `add_journal` to `create_journal`

Correct wrong bracket for reject function