Experiment with angular-cache
One prominent example where we can employ frontend-side caching
is the fetching of associated users in activities.

For a large number of comments of the same user, that user is retrieved
once per activity, which causes a drastic overhead.

This commit introduces a simple wrapper around angular-cache to store
values in the sessionStorage.