Always allow attachment upload for new work packages
https://github.com/opf/openproject/commit/af73d607733c56d1369a9d914b9ce6dd5b352f43
introduced a check whether the work package contains a link addAttachment
(i.e., the user is allowed to :edit_work_packages).

For new work packages, this link is not available, thus we need to
provide some alternative as for now, when adding work packages it is
always possible to add attachments (given the :add_work_packages permission).