quickfix for moving impediments
Impediments are not grouped by a root_id. As such they can not be sorted currently. The comit fixes the error being caused. It however does not disable to possibility to move an impediment witin a taskboard table cell (e.g. from after another impediment to before it). Those changes will be lost after a reload.