rely on fully qualified uris for HALAPI
As we should more and more rely on the links provided by the API, it makes less and less sense to construct the urls by ourselves. We can simply use the links, the api offers. There are still a lot of places that construct the uri manually. The HALAPIRessource adds to that manual construction by adding [appBasePath] + /api/v3 to every path provided to it. The callers do now have to care for that. The current callers have been altered so that they do by relying on the PathHelper which should reduce the places in which uris are constructed.