have labels for all the inplace edit fields
The labels are hidden, so they do not disturb. On the other hand, they enable blind users to have the fields properly named for them by a screen reader. It would be better, to have the element to the right of the input element (in the dt) turn into a label. But given the current structure of the directives that would have been way more complicated