Re-implement my/page block drag and drop in dragula and jQuery
We part ways with Protoype's Sortable module, link_to_remote, and some javascript
in favor of a tigher implementation using dragula, jQuery, and some js.erb
responses.

Sorting logic was also simplified; we now merely send the state of the source
and the target of the drag&drop to the server where it persists. We no longer
make a special case for each of the containers.