Fix status reporting cukes
https://github.com/opf/openproject/pull/3711 moved the status reportings
menu into the timeline module.

This in turn breaks some cukes that rely on clicking the status
reportings menu.

Also, we need to extend these cukes with an existing timeline, otherwise
a new timeline form is rendered (for which some of the cukes do not
account permissions for and again, fail).