Let templates be cached by ngtemplate-loader
Currently, the templates in this plugin do not work in installations
with relative root due to the way they are loaded (as simple assets with
an absolute path).

Instead, we can manually require the templates in the
`openproject-costs-app.js` and in turn, let `ngtemplate-loader` do the
lifting of putting the templates in $templateCache.