avoid creation of production class in spec
Defining the constant CostQuery::SqlStatement will result in ruby/rails no longer looking for it in the file system. What was intended here was to reopen the class to add mocks. But if the test runs before the class is loaded initially, it defines the class and ruby will no longer look at any other place for the original file. The class created then has only a subset of the methods the original class has resulting in test failures.