Avoid validations on work package type change
This separates updateWorkPackageForm into a refreshWorkPackageForm,
which does not check validations.

This is required for updating the form from the API for type changes,
thus allowing to receive the latest form for the given type while not
reciving any validation errors nor losing data for the intersection of
form fields.