not writable overrules having a list of values
Without the additional check, fields that have writable set to false will still be considered writable if they have a link list. The priority is an example where this occured. The API should IMO not have a list of available values when the field is not writable but that is a separate fix.