copy over symbol in custom value to customizable
We need to have access to the original symbol so that the API v3 can
post process the errors on the custom fields just like it does every
other field. When only a string is provided, the symbol :unknown is
stored which will lead to the API rendering error messages like "can't
be blank" (without the erroneous field name).