Add insecure option to managed remotes
Adds an `insecure` option to the `configuration.yml`.

This option only applies when the manages: key is set to a URL.

If the given URL uses SSL, certificate checking will be disabled.
This key is set for a packaged installation by default, since it communicates locally and
these installations may often include a snakeoil certificate.