Fix copy project
After switching to strong parameters mass assignment for project params
is not supposed to work without permitting params. We now pass a normal
hash instead of a ActionController::Params hash. Otherwise we get an
ActiveModel::ForbiddenAttributesError

All errors on the job are being rescued and logged to a separate file.
Normally rescueing all errors on Exception is not something you want to
do but as the job has an ensure block you won't know about any failures.