allow to disambiguously group by responsible association
Before the commit, the result for a count on a query grouped by
responsible would be of the schema { id: count, id2: count2 }. For other
attributes, e.g. assignee it is { user: count, user2: count2}. The
difference is caused by using :responsible_id for grouping which is
necessary as the association causes an ambiguous reference in the issued
db query. That problem has been fixed in rails/master
(https://github.com/rails/rails/commit/c2d33c4abf01ec25b2d1528a7dae923aef5a30af).
But as long as the fix is not backported, we will have to overwrite a
method of ActiveRecord::Calculations