Remove empty href in new work package menu
This removes the href attribute from the menu items, since it causes
weird behavior depending on the subproject the user is in.
It uses ui-sref for WP-create links instead.

For the user, this should not have a drawback (rather, it improves
readability for screenreaders since they do not pronounce 'Link, visited' all the time).

The downside is that capybara's `click_link` does not match without a
link. We could do `href='#'` or `href='javascript:;'`, but I dislike
both. Alternatively, we should refactor these non-link items as a button
instead, because essentially, that's what they are.