limit joined members
Prevents to have all the user's members joined on a query for a
project's available_assignees and available_responsibles. Not only the
memberships the query is interested in (the one of the project) was
joined but all the memberships the user had. So the more memberships the
users have, the slower the query did become.