Remove double escaping
With OpenProject 5.0, we moved to Open3 pseudo-shell, which no longer
requires escaping seperate arguments (unless the whole command is passed
as a single string, which will result in an actual shell that would
require escaping).

The `shell_quote` functionality was still used though and causes issues
with remote subversion passwords, which will be improperly escaped.