improve performance of projects table
The purpose of the subqueries is for them to be factored into a sum.
Therefore rows that are null and filtered out rows count the same - 0.
But performance wise it makes a lot of difference if the row needs to be
kept in the query or if it is removed early on. This commit changes the
query to no longer keep rows for which no attachment exists by turing
the LEFT JOIN into a JOIN. By that, all containers without an attachment
are ruled out very fast. If a LEFT JOIN is used, all rows for work
packages and wiki/wiki_pages are returned even if they have no
attachment.