Add separate regex-based delimiter for incoming mails
This adds a simple additional field to let the input body of mails be
delimited by a regex provided by the user, since all other delimiters
are escaped before being used by regex.

Since we don't want to force existing user of the feature to change
their settings, simply removing the regex escape is not a viable option.