Use back_url for bulk move, copy, and edit
This sets the back_url to wherever the bulk request originated from,
allowing users to be redirected to their old work package, query, or
table.

Note that this does not simply work for bulk destroy, since users may
use that functionality from split/full view and will in this case be redirected
to a non-existing resource.
We might pass a query_id in this case (to at least redirect back to queries), however it would be better if the
frontend decided what path the user should be redirected to in this
case.