eager load in decorator
Eager loads the represented work packages based on a class variable.

This might look like it would break encapsulation at first and it does. When it comes to eager loading one has to break encapsulation at some point. At least this way, we are very near to what we are going to represent and thus know which models we need.

Additionally, this enables plugins to register for being eager loaded