less magic on default query columns
Use the columns defined to be default for a new query as a default for a new query upon initialization. A caller can chose to overwrite this. This prevents the columns from getting readded when no column is selected.

It also allows to improve the performance of the work_packages#index call because an SQL query with less columns is considerably faster.