Collapse main menu early on loading the DOM
This really is a hack around the initialization of angular for the main
menu. Since the state of the menu is determined through sessionStorage,
we can leverage the immediate DOM placement to modify the menu before
the full page has been loaded.

Otherwise, the main menu flickers when collapsed (i.e., is first loaded
expanded and when the page is loaded, is being collapsed).