Inline edit: Add tooltip when form not editable
This adds a toolbar to attributes which are writable from their schemas,
but are no longer writable when their form is loaded.

This may happen for, e.g., priority of parent work packages, but also
other magic constraints.

While we work on reducing the magic in the constraints, a generic tooltip
offers the fastest user-friendly way of informing the user of
constraints.