Remove parse error when opening details view
A little refactoring also took place to satisfy
the IDE - adding type definitions.
The distinction between arrow and ordinary
functions clarifies, where the arguments
object is needed and where it is not.

The useCaching variable always returned a
truthy value, because of the `|| true` at the
end. After removing it, the application startet
to function normally.