Remove protractor E2E tests
In agreement with @furinvader, we have decided to remove protractor
testing from OpenProject.

Frontend tests are implemented in karma and test single units of work as
before.
Since the freshness of APIv3 JSON mocks are getting ouf hand and we have
covered almost all E2E protractor tests with Rails, we're going to
resort to Rails to write concise *actual* integration test with correct data.