Improve stored settings as a whole
Currently, settings are cached on a per-name basis, but that cache is
expired everytime *any* attribute is written.

Even when a week-old attribute is still valid, its cache is expired as
soon another setting is updated due to the max settings cache key.

Instead of caching a single attribute, I suggest to cache the entire
settings (the total number with plugins is less than 120 and 4KiB in
size) and retrieve and deserialize a single value from that hash.

That saves us the time to:

1. Locating a single element from database
2. Instantiate a setting object just for deserializing
3. Requesting settings several times by using RequestStore