Refactor how table grouping is applied and refreshed
Currently, changes to groupBy causes the work package to be immediately
refreshed before the actual groups are collected from the backend.
This watch has been removed, since its unneeded.

The work package table is now tracked by the row id and the current
group, which causes the ng-repeat to re-evaluate when the group changes,
thus fixing the group header bug.

Additionally, we now carry the groupByColumn in the wpTableService,
where the other grouping values are derived as well.