Refactor menu splitting and project drop down
This allows us to use the helper to decide when a drop-down node is
selected based on its content.

Since we do not always pass MenuItems to the render_drop_down_node
helper, we unfortunately must pass the selected state of its contents in that
case.

Also moves project and help helpers into separate modules