Fix/representer cache key (#4385)
Moves the calculation for whether the WP schema representer has to be rerendered to the representer itself in the form of the cache_key method. This reduces coupling as the cache_key calculation is based on information the representer already has and the controller now no longer needs to know that information.

It additionally uses the same calculation for the E-Tag to reduce duplication.

In order to allow plugins to add to that cache_key, the acts_as_op_engine module is extended.