Feature/latest project activity (#4390)
The project table is extended to also have the "Latest activity at" column which shows the date of the latest write operation within the project.

https://community.openproject.com/work_packages/22858

Models can register for being considered when the latests activity date is calculated. This mechanism can be and is used by core and can also be used by plugins.

It is unfortunate, that the mechanisms already in place by acts_as_activity could not be reused, but it would have required a complete rework to make it happen.

Another shortcoming of the current implementation are the sql strings generated within Project::Activity. It would have taken more time to rework them into Arel and as the improvable code is limited to one class and a clean interface separates it from the rest of the application I deem it ok to stay that way for now.