Fix/invalid date entry (#4398)
Allows to correct errors made when data of an invalid format is inserted to one of the WP inline edit fields, e.g. when 'aaaaa' is inserted for 'start date'.

Before the fix, the promise calling 'POST work_packages/:id/form' was assigned to the work package resource. But in case of an error, the result of the promise is an error and not a form. As we do rely on the links present in the form, we need to ensure that we always have a form where we expect it.

The PR also fixes a couple of errors in the js console when invalid information is added which are caused by the API not specifying the invalid column. While the PR increases the robustness of the front end, it does not improve the API.