add ability to activate inline edit fields via keyboard (#4405)
Allows tabbing on the inline edit field.

https://community.openproject.com/work_packages/23076/activity

I feel that one scenario is still not supported the way it should be. If the user presses the "ESC" key while being in an inline edit input, the input disappears and the text representation is displayed as desired. But the focus is not placed on the text so it becomes hard to reactivate the field as the user has to start the tabbing journey anew.

I couldn't get this to work as the field can only be focused after it reappears.