Do not store allowedValue when it is empty
If the user may edit a work package but is not a member of that project
or has no workflow (e.g., Admin without role in that project), it may
start editing a workpackage, but fail to save it due to the status being
empty, but required.

We should not copy the allowed value of a property when its embedded schema doesn't allow changing it.