unified date field for milestone work packages (#4437)
Work packages having a type that is a milestone receive a 'date'
attribute in their API resource. Other work packages still have the
'startDate'/'dueDate' attributes.

The schema behaves analogous so that such a change is visible to a
client.

In the cient we can now remove some of the custom behaviour for the date
field. The code removal is limited to the new inplace edit
implementation as the old implementation can hopefully be removed fully
in the near future.