Fix saving/display of boolean custom fields
Boolean custom fields were no longer saved since the field was marked as
required and thus invalid when unchecked.

Even when booleans were unchecked, they were output as 'Yes' since the
text was evaluated to become the placeholder `-` before the format
helper was called, thus resulting in a truthey value.

Also, unchecked booleans are marked as placeholders since they are
falsey when empty.