introduce strategy for displaying fields
In order improve the code structure and in order to enable plugins to
add display strategies for fields, that are not covered by the core yet,
a strategy pattern is introduced for displaying fields.

The mechanism is identical to the way EditFields are handled. In fact,
the DisplayFields and EditFields now inherit from a common superclass,
Field. Also, the service and the factory inherit from a common
superclass.