Do not pass empty values to smtp configuration
Even when specifying `SMTP_AUTHENTICATION=none`, SMTP doesn't really have such an authentication method.
Instead, it looks for presence of `user_name` and `password` and when found, assumes an authentication method to be available.

It then fails to find the auth method `none`.

This is a regression of the configurable mail settings. With settings,
default values are empty strings instead of nil values coming from
missing or empty configuration / ENV variables.

Checking the values for presence is sufficient to resolve the issue.