Fix summing done_ratio when estimated_hours is zero (#4528)
When `estimated_hours` is `0.0`, it is still taken into consideration
when weighting average done ratios for a parent. That results in the
parent becoming `0%` done.

As 0.0 is being set in the frontend as the default value, this is now
happening all over the board.

A separate PR should refactor the frontend for 5.1. to allow null as a
value for estimated_hours.