Fix/accessible error messages (#4529)
* add error messages to each field

* move role=alert up so that success messages are also covered

* remove focus code from notification-box directive

The box does not need to be focused as it has the role=alert which will lead to it being read out.

* remove specific focusing behaviour for accessibility mode

The mode was introduced to present error messages to user requiring assistive technology e.g. screen readers. By setting role=alert on the error fields and additionally labeling the fields with the errors, this should no longer be necessary