Fix/accessible pre and suffix (#4601)
* accessible form pre- and suffix

Both pre- and suffix are attributes with aria-hidden="true". This should
lead to them no longer being read out of context.

In order to have them read when appropriate, the prefix is turned into a part
of the label which is hidden for sighted users. The suffix is referenced
in the form element via aria-describedby.

* extract method refactoring