Remove attachment service/directive coupling
Previously, attachments for a work package were kept in the directive
(good, since its scoped by the active work package and destroyed when
 moving) AND in the service (horrible, since they are requested twice
AND the service is never reset).

This causes a bug when moving to new/existing work packages due to
remaining attachments in the service.

We can completely avoid this by passing an event from drag handler to
the directive instead to the service.

This also removes the edit / no-edit attachment template duality, since
that's no longer required.