Bugfix for creating work packages globally
supports deep links so that one can call work_packages/new
fixes opening the work package form when in a global context. The type is set to null when in that context which broke the controller. https://community.openproject.com/work_packages/23595/activity
harmonises the usage of a border for show/edit and create split view
avoids a url of [domain]///work_packages/new when switching from split to full view in a global context
Out of scope

adding 403 error messages when a request to the API by the angular front end fails with that error. Currently, the content remains blank which does not reveal any information. https://community.openproject.com/work_packages/23615/activity