Use contextReplacementPlugin to reduce loaded ngLocales
A dynamic require, such as the following in `init-app.js` will cause
webpack to include all possible values (since it doesn't know what
`documentLang` is in this case).

```
// depends on the html element having a 'lang' attribute
var documentLang = (angular.element('html').attr('lang') || 'en').toLowerCase();
require('angular-i18n/angular-locale_' + documentLang + '.js');
```

We can exploit the `webpack.contextReplacementPlugin` to match this
context and restrict the possible modules that may match it.

Applying that for ngLocale reduces the number of included locales to the
ones we have in place in the `openproject-translations` plugin, and all
other locales are excluded. That number drops from 750 locales to 34,
which results in a drop of 1.9 MB in the openproject-core bundle.