adhere to notification settings when notifying about added watcher (#4683)
The PR will let the users added as watchers be notified according to their settings. Especially https://community.openproject.com/work_packages/22636/activity should be fixed by it.

Unfortunately, the #notify_about method on User is not reusable here for all cases as it would have to be adjusted to fit the requirements. But that method is used for a couple of use cases and I did not want to alter it shortly before a release. The method looks weird, though as e.g. the watchers are not taken into account for 'only_my_events' which reads 'Only for things I watch or I am involved in'. Additionally, it does not care for the self_notified? setting.