[23697] Remove project-specific version from representer
Grouping versions are performed on their names, not on the
project-specific name (e.g., `parent - version` in a subproject where
this version is shared with).

This causes the group name and row value to differ and cause an error.

This commit changes the behavior as before. Grouping instead should be
made on a unique identifier (valueLink for non-scalar values) and the
grouping should generate a project-specific output so that versions can
add the above project-specific name.