[23753] Use MenuItem#title to discover the page
Since WikiMenuItem doesn't provide any other locator other than the
title, we can employ `wiki.find_page` to locate the page by title.
That title is either matched to the slug, or found by the legacy
redirect introduced with 6.0.1.

This still feels weird, but without an explicit column or accessor to
the page, there's not much we can do while keeping compatibility.

https://community.openproject.com/projects/openproject/work_packages/23753