Remove absolute footer from work package layout
The inner scrolling combined with the increased complexity of the table
(cells) causes FireFox to go nuts when 200+ work packages are showing.
The work package footer (pagination area) jumps up and down when
resizing and hovering in the table. This is likely caused by a repaint
that initially collapses the height of the table determined by
flex-grow.

As the height is actually always fixed (full height minus bottom
pagination), we don't need absolute positions and flex-based heights.