[23715] Keep current project context as stateParam
This adds a global projectPath parameter to the router configuration,
which allows us to keep the current project context without explicitly
specifying it as a parameter.

We do not have frontend movements outside the project context without a
hard refresh. Should that come, one simply has to explictly specify
`projectPath: null` to the route request.