hide date picker while scrolling and repaint
As the date picker is positioned absolutely and we can not guarantee to
position all scrolling containers so that they allow for scrolling the
absolutely positioned element along with the scrolling container
(relative), we hide the date picker as long as the users scrolls. Upon
being finished with scrolling, wre repaint the date picker again.