[23859] Respect intermittent additions to WP pristine
When a user quickly saves a first attribute and opens a second attribute
before the first request has completed, the `$pristine` value stored for
the second is removed after the API request has completed and thus, the
second change is not saved.

A workaround is to remove only those pristine values that were known
before saving the resource.