[23853] Post the original form attributes when copying
Previously, the copied work package would take over all attributes of
the original work package, but the form was an empty create form, which
results in only the default type and status being contained in the
schema.

We can instead post the attributes of the original WP form to retrieve a
matching form response.