normalize permissions
With rails 5 it is no longer possible to create an sql like statement via arel which will check whether a serialized array contains a string. AR will only allow to create a match against an array if the column itself is marked to serialize an array. Because of this and in order to increase the normalization of the data, the permissions are no longer serialized and are  moved into a separate table instead. While this is still not normalized, it satisfies the needs