lazy load query filters and extract them into classes
This allows us to only load the values we actually need instead of
loading all the values regardless of whether we need them. This is
especially usefull in the api where we currently never need those
values.

Note, that this is a glitch in my opinion as we do not check whether the
filter value is valid.