[23896] Submit and close edit mode on enter for most edit fields
When editing a single field (except select, checkbox), pressing enter
triggers the `submit` handler of the surrounding form.

This form submit is disabled when in edit mode, since the wp-edit-field
uses that submit also for saving on blur, which we deliberately disable
in this case.

We can instead capture another event on those fields where it makes
sense that submits the field after pressing enter, but only IN edit mode
(since without it, submit handles it).

https://community.openproject.com/work_packages/23896