Change specification: Only admins are permitted
When user's should be allowed to update their own parameters,
we will need extra care to not interfere with the password confirmation
requirements.

As there is no explicit requirement that this should be possible, I'm
restricting the specification to allow only admins to update users
without confirmation, identically to how the admin UI allows that.