Implement interval checking for correct table and header widths
Multiple bugs exist when parts of the table are being resized instead
of the entire window.

Currently, after first initialization, the interactive-table updates itself when:

1. The window is being resized (resize listener)

2. A state change has successfully transitioned (e.g., split view  opened)

3. The project menu is being toggled

In some cases, these listeners will be execeuted when the content of the
table is not yet ready. One prominent example is the initial loading of
the table when, e.g., the progress column is active and partly out of
the window. It will be cut off and not redrawn.

When excluding the option of bottom-up notification on when a cell or
row changes in the table (pure chaos), the only way to address this is
to periodically check the table and header width for changes and update
the table accordingly.

This is exactly what this commit suggests.