[24083] Don't show the drop-down until the form and types are loaded
Currently, expanding the types drop down will initiate the form load and
shows a small part of the drop down despite it being empty.

Possible solutions:

1. Preload the form. Doesn't necessary fix the issue if the user is fast
enough and puts more initial effort on the page

2. Don't show the dropdown until types are loaded. Here we can't remove
it from the dom with `display: none` since it will mess with the
positioning, so we only use `visibility: hidden` to render, but not show
it.

https://community.openproject.com/work_packages/24083