Enhance delayed_job to load full environment (including eager load)
We use delayed_job's `jobs:work` rake task in packaged environments to
run the worker process.

As it turns out, Rails doesn't eager load on rake tasks by default:
https://github.com/rails/rails/blob/master/railties/lib/rails/application.rb#L451

But without eager_load and autoload, not all dependencies may be met and
this causes a nasty invalid constant in the documents plugin, where
`Document` becomes an auto-record from the table information but the
`app/models/document.rb` requirement has never been loaded.

The suggested solution is to load the full environment in a separate
rake task and point the prerequisites of `jobs:work` towards that.

[ci skip]