Fix relation caching issue
This is a hack. Relation links are being cached with the CacheService
(as are all GET requests unless explicitly stated otherwise).

Now if you force to reload a work package, it does not reset its links
and thus those are still cached.

There's no quick fix for this, but the correct solution is to rewrite
the caching using states.