[24233] Fix deep link into work package project context
When deep linking to a work package such as `/work_packages/:id`,
Rails renders a project context (project menu, etc.), but the frontend
does not respect the `projectIdentifier` setting as long as the
`projectPath` is missing from ui-router.

We can enforce the projectPath by looking at the root projectIdentifier.
If that is set, we're within the project context.