[24258] Reuse existing schema after updating work package
When a work package is updated after a successful PATCH, its schema is
being reset to the linked value (and thus, `$loaded == false`).

Since this controls type attribute visibility, this always leads to a
small period of time where the schema is being retrieved from cache and
causes a visible reloading of an attribute section.

Here, again, the correct fix is to AVOID loading schemas as links, but
rather as a separate state where the referenced schema will always
reside unless the switch was made to a new type.