[24262] Load broken queries to let the user fix it
When the API experimental returns a query that turns out to be invalid,
an error is returned from APIv3. This causes the filters to remain
hidden and thus leaves the user unable to correct the query.

Instead, we can render the query with a pseudo-empty response and create
a warning message similar to the no results row.