Use webpack.optimize plugins for test and production
This applies uglifyjs with compression (no name mangling) and other
webpack.optimize plugins anytime the RAILS_ENV is not development.

I've tried this in the past but received errors due to uglifyjs causing
the html-loader to also minify html, causing all sorts of errors.
However, we can separately disable this minification.

This cuts the size output bundles by around half.