fix estimated time filters
Because estimatedTime is turned to estimatated_hours by the PropertyNameConverter, checking whether work package responds to estimated_hours.pluralize only to return estimated_hours_id does not make sense. That behaviour was introduced to cope with the watcher attribute which should have been turned into watcher_ids as watchers is a has_many relations.
But when that behaviour was fixed to work correctly with estimated_hours again, so that the name converter returns estimated_hours (no foreign key) and watcher_ids (foreign key) the watcher bug had to be fixed all over again.
For turning work package filter parameters from the external to the internal representation, the v3 now makes use of the same method, as employed by the experimental api