prevent WorkPackage#visible to result in duplicate entries for pluck
If Project.allowed_to is joined to WorkPackage, the result set will have as many entries, as the user has roles (with the queried for permission) in a given project. Normally, this is not a problem as rails prevents instantiating duplicate WorkPackages on a single call. But when we only pluck, then, in combination with limit, we will end up with too few results as we get duplicate rows and only take the first x once.