Separate actual documentation guides in github from sources
From now on, use `doc/` as a source folder for documentation assets to
be built (e.g., API documentation or source files included in wordpress).

Use `guides/` for self-reliant Markdown documents on topics relevant for
developers and administrators, as well as community-contributed guides
we do not feature on the website.

[ci skip]