Add custom sub-models for attribute constraints and groups
This allows plugins such as backlogs and costs to integrate
into the grouping while restricting while they are visible.

The problem here is that on a global admin page, the plugin's attributes
should be available, however in a project context, they should only
be added when the respective module is active.