Fix CF translations with non-matching locales
While `globalize` provides fallbacks, we allow custom fields with _any_
set of locales being translated. So it could be that a CF exists with
only a german (or any other locale) translation.

Globalize will not automatically take the first available translation as
a fallback, so we have to implement that ourselves:

1. Remember the first saved translation for any CF (defined by its
        order, which matches the appearance of fields).

2. Use the globalize's own `drop_translation_table!` helper to move back
the current locale.

3. For any CFs with their name being `null` now, update it with the
first available value.