[25096] Fix initial positioning of dropdown
The default position of the dropdown right before positioning caused it
to be inside the normal dom flow, resulting in a temporary scroll bar,
which shifted the total width of the container.

If we instead set a default top/left value which is immediately cleared
on the first reposition, the position is correct.