Remove template rendering in wp-display-fields
This fixes a recurring compiling of a template in a new scope that is
never cleared.

The SimpleTemplateRenderer is still in use for the edit field
functionality in the table, however there we have an explicit lifetime
that can be controlled through the EditFieldHandler.

Thus, it keeps a reference to the scope used to compile the field, and
destroys it upon deactivation.